CLOUD-2719, CLOUD-2512, CLOUD-2698, CLOUD-2601

+++ migrations/versions/98c3d8971d6d_second_update.py
@@ -6,7 +6,7 @@ Create Date: 2020-03-04
 """
 from alembic import op
 import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
+from app.lib import alembic_helper
 
 # revision identifiers, used by Alembic.
 revision = '98c3d8971d6d'
@@ -16,12 +16,16 @@ depends_on = None
 
 
 def upgrade():
-    op.add_column('deployments', sa.Column('updatable', sa.Boolean, server_default=sa.sql.False_(), nullable=False))
-    op.add_column('deployments', sa.Column('keep_last_attempt', sa.Boolean, server_default=sa.sql.False_(), nullable=False))
+    if not alembic_helper.column_exists('deployments', 'updatable'):
+        op.add_column('deployments', sa.Column('updatable', sa.Boolean, server_default=sa.sql.False_(), nullable=False))
+    if not alembic_helper.column_exists('deployments', 'keep_last_attempt'):
+        op.add_column('deployments', sa.Column('keep_last_attempt', sa.Boolean, server_default=sa.sql.False_(), nullable=False))
     # ### end Alembic commands ###
 
 
 def downgrade():
-    op.drop_column('deployments', 'updatable')
-    op.drop_column('deployments', 'keep_last_attempt')
+    if alembic_helper.column_exists('deployments', 'updatable'):
+        op.drop_column('deployments', 'updatable')
+    if alembic_helper.column_exists('deployments', 'keep_last_attempt'):
+        op.drop_column('deployments', 'keep_last_attempt')
     # ### end Alembic commands ###
