+++ migrations/versions/98c3d8971d6e_third_update.py
@@ -6,7 +6,7 @@ Create Date: 2020-03-22
 """
 from alembic import op
 import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
+from app.lib import alembic_helper
 
 # revision identifiers, used by Alembic.
 revision = '98c3d8971d6e'
@@ -16,10 +16,12 @@ depends_on = None
 
 
 def upgrade():
-    op.add_column('deployments', sa.Column('stinputs', sa.Text, nullable=True))
+    if not alembic_helper.column_exists('deployments', 'stinputs'):
+        op.add_column('deployments', sa.Column('stinputs', sa.Text, nullable=True))
     # ### end Alembic commands ###
 
 
 def downgrade():
-    op.drop_column('deployments', 'stinputs')
+    if alembic_helper.column_exists('deployments', 'stinputs'):
+        op.drop_column('deployments', 'stinputs')
     # ### end Alembic commands ###
