+++ migrations/versions/98c3d8971d6f_fourth_update.py
@@ -6,7 +6,7 @@ Create Date: 2020-04-02
 """
 from alembic import op
 import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
+from app.lib import alembic_helper
 
 # revision identifiers, used by Alembic.
 revision = '98c3d8971d6f'
@@ -16,10 +16,12 @@ depends_on = None
 
 
 def upgrade():
-    op.add_column('deployments', sa.Column('selected_template', sa.Text, nullable=True))
+    if alembic_helper.column_exists('deployments', 'selected_template'):
+        op.add_column('deployments', sa.Column('selected_template', sa.Text, nullable=True))
     # ### end Alembic commands ###
 
 
 def downgrade():
-    op.drop_column('deployments', 'selected_template')
+    if alembic_helper.column_exists('deployments', 'selected_template'):
+        op.drop_column('deployments', 'selected_template')
     # ### end Alembic commands ###
